requirements.txt
aiosqlite==0.18.0
alembic==1.10.2
anyio==3.6.2
apprise==1.3.0
asgi-lifespan==2.1.0
asyncpg==0.27.0
attrs==19.3.0
Automat==0.8.0
backports.zoneinfo==0.2.1
blinker==1.4
cachetools==5.3.0
certifi==2019.11.28
cffi==1.15.1
chardet==3.0.4
charset-normalizer==3.1.0
click==8.1.3
# cloud-init==23.1.1
cloudpickle==2.2.1
colorama==0.4.3
command-not-found==0.3
configobj==5.0.6
constantly==15.1.0
coolname==2.2.0
cramjam==2.6.2
croniter==1.3.8
cryptography==40.0.1
dateparser==1.1.8
dbus-python==1.2.16
distro==1.4.0
distro-info===0.23ubuntu1
docker==6.0.1
entrypoints==0.3
fastapi==0.95.0
fastparquet==2023.2.0
fsspec==2023.1.0
google-api-core==2.11.0
google-api-python-client==2.83.0
google-auth==2.17.1
google-auth-httplib2==0.1.0
google-cloud-bigquery==3.9.0
google-cloud-bigquery-storage==2.19.1
google-cloud-core==2.3.2
google-cloud-storage==2.8.0
google-crc32c==1.5.0
google-resumable-media==2.4.1
googleapis-common-protos==1.59.0
greenlet==2.0.2
griffe==0.25.5
grpcio==1.53.0
grpcio-status==1.53.0
h11==0.14.0
h2==4.1.0
hpack==4.0.0
httpcore==0.16.3
httplib2==0.22.0
httpx==0.23.3
hyperframe==6.0.1
hyperlink==19.0.0
idna==2.8
importlib-metadata==6.1.0
importlib-resources==5.12.0
incremental==16.10.1
Jinja2==3.1.2
jsonpatch==1.32
jsonpointer==2.0
jsonschema==3.2.0
kaggle==1.5.13
keyring==18.0.1
kubernetes==26.1.0
language-selector==0.1
launchpadlib==1.10.13
lazr.restfulclient==0.14.2
lazr.uri==1.0.3
Mako==1.2.4
Markdown==3.4.3
markdown-it-py==2.2.0
MarkupSafe==2.1.2
mdurl==0.1.2
more-itertools==4.2.0
netifaces==0.10.4
numpy==1.24.2
oauthlib==3.1.0
orjson==3.8.9
packaging==23.0
pandas==1.5.3
pathspec==0.11.1
pendulum==2.1.2
pexpect==4.6.0
prefect==2.9.0
prefect-gcp==0.3.0
proto-plus==1.22.2
protobuf==4.22.1
pyasn1==0.4.2
pyasn1-modules==0.2.1
pycparser==2.21
pydantic==1.10.7
Pygments==2.14.0
PyGObject==3.36.0
PyHamcrest==1.9.0
PyJWT==1.7.1
pymacaroons==0.13.0
PyNaCl==1.3.0
pyOpenSSL==23.1.1
pyparsing==3.0.9
pyrsistent==0.15.5
pyserial==3.4
python-apt==2.0.1+ubuntu0.20.4.1
python-dateutil==2.8.2
python-debian===0.1.36ubuntu1
python-slugify==8.0.1
pytz==2023.3
pytz-deprecation-shim==0.1.0.post0
pytzdata==2020.1
PyYAML==6.0
readchar==4.0.5
regex==2023.3.23
requests==2.28.2
requests-oauthlib==1.3.1
requests-unixsocket==0.2.0
rfc3986==1.5.0
rich==13.3.3
rsa==4.9
SecretStorage==2.3.1
service-identity==18.1.0
simplejson==3.16.0
six==1.14.0
sniffio==1.3.0
sos==4.4
SQLAlchemy==1.4.47
ssh-import-id==5.10
starlette==0.26.1
systemd-python==234
text-unidecode==1.3
toml==0.10.2
tqdm==4.65.0
Twisted==18.9.0
typer==0.7.0
typing_extensions==4.5.0
tzdata==2023.3
tzlocal==4.3
ubuntu-advantage-tools==8001
ufw==0.36
unattended-upgrades==0.1
uritemplate==4.1.1
urllib3==1.26.15
uvicorn==0.21.1
wadllib==1.3.3
websocket-client==1.5.1
websockets==11.0
zipp==3.15.0
zope.interface==4.7.1
